consumeGold 메소드 삭제 및 이동
