+++ src/main/java/org/newdawn/spaceinvaders/User.java
@@ -1,5 +1,13 @@
 package org.newdawn.spaceinvaders;
 
+import com.google.firebase.database.DatabaseError;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
+
+import java.util.Base64;
+import java.util.HashMap;
+import java.util.Map;
+
 public class User {
     private String email;
     private String password;
@@ -49,6 +57,21 @@ public class User {
 
     public void setGold(Integer gold) {
         this.gold = gold;
+
+        String encodedEmail = Base64.getEncoder().encodeToString(getEmail().getBytes());
+        FirebaseDatabase database = FirebaseDatabase.getInstance();
+        DatabaseReference ref = database.getReference();
+        DatabaseReference userRef = ref.child("Users").child(encodedEmail);
+
+        Map<String, Object> updates = new HashMap<>();
+        updates.put("/" + encodedEmail + "/gold", getGold());
+
+        userRef.updateChildren(updates, new DatabaseReference.CompletionListener() {
+            @Override
+            public void onComplete(DatabaseError databaseError, DatabaseReference databaseReference) {
+            }
+        });
+
     }
 
     public String getBestTime() {
